Metric Tests

+++ flight/test/test.cpp
@@ -197,8 +197,8 @@ TEST(MetricTests, MetricTestOurs){
   string fuel_burn = "";
   getline(output, response);
   getline(output, fuel_burn);
-  float res_time = atof(response.substr(response.length() - 2));
-  float fuel_avg = atof(fuel_burn.substr(fuel_burn.length() - 2));
+  float res_time = stof(response.substr(response.length() - 2));
+  float fuel_avg = stof(fuel_burn.substr(fuel_burn.length() - 2));
 
   EXPECT_NEAR(4, res_time, 0.1);
   EXPECT_NEAR(21, fuel_avg, 0.1);
@@ -221,8 +221,8 @@ TEST(MetricTests, MetricTestFIFO){
   string fuel_burn = "";
   getline(output, response);
   getline(output, fuel_burn);
-  float res_time = atof(response.substr(response.length() - 2));
-  float fuel_avg = atof(fuel_burn.substr(fuel_burn.length() - 2));
+  float res_time = stof(response.substr(response.length() - 2));
+  float fuel_avg = stof(fuel_burn.substr(fuel_burn.length() - 2));
 
   EXPECT_NEAR(7, res_time, 0.1);
   EXPECT_NEAR(18, fuel_avg, 0.1);
